Added sample react-redux structure, move storybook directory.
