+++ src/containers/index.js
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux'
+
+import App from '../components'
+
+function mapStateToProps(state) {
+  return state
+}
+
+export default connect(mapStateToProps)(App)
