+++ src/reducers/index.js
@@ -1,0 +1,10 @@
+const initialState = {
+  var1: 1
+}
+
+export default function reducer(state = initialState, action) {
+  switch(action.type) {
+    default:
+      return state
+  }
+}
